Add Phase 3.6 — Re-scope analysis from account to client level + Account Overview page
Foundational re-scoping. Every analysis flow that previously ran at the
account level now requires (account_id, client_id). The account-level
endpoints return 410 Gone with code MOVED_TO_CLIENT_SCOPE, and the dashboard
moves to /accounts/[accountId]/clients/[clientId]/analysis. The bare
/accounts/[accountId] becomes the Account Overview where the user picks a
client.

Database (supabase/migrations/20260429233224_phase36_client_scoped_analysis.sql):
- portfolio_analyses: backfill nulls, ALTER COLUMN client_id SET NOT NULL,
  drop old account_module index in favor of (account_id, client_id, module_key, created_at DESC)
- account_operational_constraints: backfill, swap PK from (account_id) to
  (account_id, client_id) inside a DO $$ guard so re-runs are idempotent
- analysis_scenarios: backfill + NOT NULL + new (account_id, client_id, created_at) index
- analysis_chat_messages: ADD COLUMN client_id, backfill, NOT NULL, new composite index
- Backfill strategy: most-common client_id from this account's
  service_locations, fallback to any client of the account; orphan rows are
  deleted (none in production)

Backend:
- New api/_lib/analysis/scope.ts — resolveScope(req, db) validates that the
  client belongs to the account, returns canonical 410 body for old routes
- Shared helpers tightened to require clientId: loadConstraints,
  loadAccountProperties, loadLatestModuleSnapshots, loadAccountOfferings,
  fetchLatestCompletedAnalysis, fetchLatestBranchSet, runAllModules,
  triggerSynthesisRefresh
- All 13 analyses + 6 accounts endpoints copied to clients/[clientId]/...
  variants and updated to: read clientId from req.query, pass it through to
  every helper, set client_id on every insert, add .eq('client_id', clientId)
  to every portfolio_analyses / analysis_chat_messages / analysis_scenarios
  query, change upserts to onConflict: 'account_id,client_id'
- Old endpoints replaced with 7-line 410 Gone shims that import
  MOVED_TO_CLIENT_SCOPE from scope.ts
- New GET /api/accounts/[accountId]/overview — aggregate stats + per-client
  summary (property_count, sl_count, states_count, last_analysis_at,
  synthesis_status, branch_selection)

Frontend:
- src/App.tsx routes:
  /accounts/:accountId/clients/:clientId/analysis → dashboard
  /accounts/:accountId/analysis → AnalysisRedirectToOverview which navigates
  to /accounts/:accountId?from=legacy-analysis
- AccountDetail (now serves as the Account Overview): fetches /overview,
  shows aggregate stats, replaces the static client table with rich
  ClientOverviewRow cards that include synthesis status badge +
  "View Analysis →" / "Start Analysis →" buttons targeting the new
  per-client URL. Banner shown when bounced from legacy /analysis URL.
- AccountAnalysis dashboard reads clientId from URL params, passes it to
  every API call and every panel component, gates panel rendering on
  both accountId and clientId, and shows a breadcrumb (Accounts › JLL ›
  Property Reserve › Analysis) plus a back link
- All five panels (OperationalConstraintsPanel, ScenarioPanel, ChatPanel,
  SynthesisCard, CostAssumptionsPanel) take clientId prop and use it in
  every URL

Migration outcome: JLL/Property Reserve carries forward all 16 existing
analyses + 1 constraint row + 0 scenarios + 0 chat messages. JLL's second
client (All JLL IFM, no service_locations yet) appears as a separate row
in Account Overview with "Start Analysis →" — independent slate.
